Fix missing "replace_manager_ip.sh" script
The wazuh-agent package currently fails to upgrade from ossec-hids-agent
(and possibly others) because the "replace_manager_ip.sh" script is
missing from the /var/ossec/tmp/src/init directory. The Debian postinst
script attempts to call it but it's not there, so the postinst script
exits with an error status.