deb package: provide a short-description
The description field of a debian package control file is consisted
of two parts: the synopsis or the short description, and the
(optional) long description. The description we had was too long to be
a short description, and is now placed in the long description field.
A short description was provided.

Closes: #60