Fixes #70 for DEB packages
This commit is based in the _Upgrading_ graph from https://wiki.debian.org/MaintainerScripts. When a deb package is upgraded, the postinst script from the new package is called as follows: `postinst configure previous_version`. That's why I check if $2 is not empty.

If this is true, $2 value will be, for example, 3.6.1-1. So the package will extract the MAJOR and MINOR value from $2.