Remove unnecessary files when uninstalling the package
This commit will remove all the unnecessary files from /var/ossec/queue (sockets or unused files), databases and stats (only if you uninstall the package), etc. The files that will survive if you uninstall the package will be the configuration files.

In addition, now the package remove the .service files for systemd when the package is removed.