Small fixes to HP-UX generate package script
When exporting PATH we moved the previously defined env PATH to the end
to favor the dependencies installed by the script over the pre installed by the system.

The clean() function no longer deletes the build folder containing gcc and cmake, it only
removes the folder when the -e option is passed to the script.