Fix user and group deletion during upgrade/downgrade
In order to succesfuly downgrade to earlier versions of Wazuh, we need
to restore ownership of files to the ossec group and user.

RPM may also backup the client.keys file, making it impossible for the
agent to reconnect, for this reason we try to restore this backup of
the client.keys if none is found.